changed the restart capability to be more user friendly
